livegen/pyproject.toml
[project]
name = "livegen"
version = "0.1.0"
description = "AI-powered customer service voice agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "livekit-agents[google,nvidia,silero,turn-detector]>=1.4.3",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "python-dotenv>=1.2.1",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "prometheus-client>=0.21.0",
    "redis>=5.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
    "httpx>=0.28.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
